Move source of ansibulled to a subdir.
pyre prefers to be given the directory which holds your python package
as the source-directory rather than the directory which is your python
package.  However, if the directory given is the toplevel, pyre tends to
get slow for me.  (I believe since it scans  temporary directories that
I have accumulated [downloads of all those ansible collections...]).
Putting the soure code in its own subdirectory is the way to fix that.